Move LSDF approach2 files to own directory
