+++ src/internal/lsdf/approach2/sample.frag.glsl
